+++ plugins/ci/pkg/ci/ci.go
@@ -16,8 +16,8 @@ import (
 	"strings"
 
 	trivymodels "github.com/fairwindsops/insights-plugins/plugins/trivy/pkg/models"
+	"github.com/samber/lo"
 	"github.com/sirupsen/logrus"
-	"github.com/thoas/go-funk"
 	"gopkg.in/yaml.v3"
 
 	"github.com/fairwindsops/insights-plugins/plugins/ci/pkg/commands"
@@ -45,6 +45,14 @@ type CIScan struct {
 	repoBaseFolder string // . or /app/repository/{repoName}
 	configFolder   string
 	config         *models.Configuration
+}
+
+type insightsReportConfig struct {
+	EnabledOnAutoDiscovery *bool
+}
+
+type insightsReportsConfig struct {
+	AutoScan map[string]insightsReportConfig
 }
 
 // Create a new CI instance based on flag cloneRepo
@@ -75,6 +83,8 @@ func NewCIScan() (*CIScan, error) {
 		configFolder:   configFolder,
 		config:         config,
 	}
+
+	logrus.Infof("Reports config is opa: %v, polaris: %v, pluto: %v, trivy: %v, tfsec: %v", ci.OPAEnabled(), ci.PolarisEnabled(), ci.PlutoEnabled(), ci.TrivyEnabled(), ci.TerraformEnabled())
 
 	return &ci, nil
 }
@@ -149,9 +159,9 @@ func (ci *CIScan) getAllResources() ([]trivymodels.Image, []models.Resource, err
 						Namespace: namespace,
 						Filename:  displayFilename,
 						HelmName:  helmName,
-						Containers: funk.Map(containers, func(c models.Container) string {
+						Containers: lo.Map(containers, func(c models.Container, _ int) string {
 							return c.Name
-						}).([]string),
+						}),
 					})
 				}
 			} else {
@@ -172,9 +182,9 @@ func (ci *CIScan) getAllResources() ([]trivymodels.Image, []models.Resource, err
 					Namespace: namespace,
 					Filename:  displayFilename,
 					HelmName:  helmName,
-					Containers: funk.Map(containers, func(c models.Container) string {
+					Containers: lo.Map(containers, func(c models.Container, _ int) string {
 						return c.Name
-					}).([]string),
+					}),
 				})
 			}
 		}
@@ -224,7 +234,7 @@ func processYamlNode(yamlNode map[string]interface{}) ([]trivymodels.Image, []mo
 	}
 	podSpec := getPodSpec(yamlNode)
 	images := getImages(podSpec.(map[string]interface{}))
-	return funk.Map(images, func(c models.Container) trivymodels.Image {
+	return lo.Map(images, func(c models.Container, _ int) trivymodels.Image {
 		return trivymodels.Image{
 			Name: c.Image,
 			Owner: trivymodels.Resource{
@@ -233,7 +243,7 @@ func processYamlNode(yamlNode map[string]interface{}) ([]trivymodels.Image, []mo
 				Name:      owner.Name,
 			},
 		}
-	}).([]trivymodels.Image), images
+	}), images
 }
 
 // getPodSpec looks inside arbitrary YAML for a PodSpec
@@ -454,6 +464,14 @@ func getConfigurationForClonedRepo() (string, string, *models.Configuration, err
 			return "", "", nil, err
 		}
 
+		// this is how we support enabling/disabling reports on auto-discovery (when no fairwinds-insights.yaml file is found)
+		if strings.TrimSpace(os.Getenv("REPORTS_CONFIG")) != "" {
+			err := unmarshalAndOverrideConfig(config)
+			if err != nil {
+				return "", "", nil, err
+			}
+		}
+
 		err := createFileFromConfig(baseRepoPath, configFileName, *config)
 		if err != nil {
 			return "", "", nil, err
@@ -480,6 +498,34 @@ func getConfigurationForClonedRepo() (string, string, *models.Configuration, err
 	}
 
 	return filepath.Join(baseRepoPath, "../"), baseRepoPath, config, nil
+}
+
+func unmarshalAndOverrideConfig(config *models.Configuration) error {
+	var insightsReportConfig insightsReportsConfig
+	err := json.Unmarshal([]byte(os.Getenv("REPORTS_CONFIG")), &insightsReportConfig)
+	if err != nil {
+		return fmt.Errorf("unable to parse auto-scan reports config: %v", err)
+	}
+	overrideReportsEnabled(config, insightsReportConfig)
+	return nil
+}
+
+func overrideReportsEnabled(cfg *models.Configuration, reportConfig insightsReportsConfig) {
+	if rCfg, ok := reportConfig.AutoScan["opa"]; ok {
+		cfg.Reports.OPA.Enabled = rCfg.EnabledOnAutoDiscovery
+	}
+	if rCfg, ok := reportConfig.AutoScan["polaris"]; ok {
+		cfg.Reports.Polaris.Enabled = rCfg.EnabledOnAutoDiscovery
+	}
+	if rCfg, ok := reportConfig.AutoScan["pluto"]; ok {
+		cfg.Reports.Pluto.Enabled = rCfg.EnabledOnAutoDiscovery
+	}
+	if rCfg, ok := reportConfig.AutoScan["trivy"]; ok {
+		cfg.Reports.Trivy.Enabled = rCfg.EnabledOnAutoDiscovery
+	}
+	if rCfg, ok := reportConfig.AutoScan["tfsec"]; ok {
+		cfg.Reports.TFSec.Enabled = rCfg.EnabledOnAutoDiscovery
+	}
 }
 
 func readConfigurationFromFile(configFilePath string) (*models.Configuration, error) {
@@ -613,7 +659,7 @@ func (ci *CIScan) ProcessRepository() ([]*models.ReportInfo, error) {
 	}
 
 	if ci.TerraformEnabled() {
-		terraformReports, err := ci.ProcessTerraformPaths()
+		terraformReports, areTerraformResults, err := ci.ProcessTerraformPaths()
 		if err != nil {
 			scanErrorsReportProperties.AddScanErrorsReportResultFromError(models.ScanErrorsReportResult{
 				ErrorMessage: err.Error(),
@@ -621,7 +667,9 @@ func (ci *CIScan) ProcessRepository() ([]*models.ReportInfo, error) {
 				Kind:         "InternalOperation",
 				ResourceName: "ProcessTerraformPaths",
 			})
-		} else {
+		}
+		if areTerraformResults {
+			logrus.Debugln("the Terraform report contains results")
 			reports = append(reports, &terraformReports)
 		}
 	}
@@ -690,6 +738,14 @@ func (ci *CIScan) printScannedFilesInfo() {
 		fmt.Println("Helm charts scanned:")
 		for i, h := range ci.config.Manifests.Helm {
 			fmt.Printf("\t[%d/%d] - %s/%s\n", i+1, s, h.Path, h.Name)
+		}
+	}
+
+	s = len(ci.config.Terraform.Paths)
+	if s > 0 {
+		fmt.Println("Terraform files scanned:")
+		for i, p := range ci.config.Terraform.Paths {
+			fmt.Printf("\t[%d/%d] - %s\n", i+1, s, p)
 		}
 	}
 }