+++ plugins/ci/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 4.2.0
+* CI scanning will continue when an error is encountered, such as templating a Helm chart into Kubernetes manifests. These errors will be reflected as Insights action items, in a new `ScanErrors` report type.
+
+## 4.1.0
+* Add support for configuring reports when using auto-discovery via `REPORTS_CONFIG` env var
+
+## 4.0.0
+* Enable the tfsec report by default. If `terraform -> paths` are specified, they will be scanned unless `reports -> tfsec -> enabled` is explicitly set to `false` in fairwinds-insights.yaml.
+
 ## 3.4.0
 * Support for private images (REGISTRY_CREDENTIALS)